# static/openapi.yaml
openapi: 3.0.3
info:
  title: CardioVoice API
  version: "1.0.0"
  description: |
    CardioVoice Backend API - Health risk assessment using voice analysis and AI-powered recommendations.
    
    ## Authentication
    This API uses JWT Bearer token authentication. To access protected endpoints:
    1. Register a new account via `/api/v1/auth/register`
    2. Login via `/api/v1/auth/login` to receive an access token
    3. Include the token in the `Authorization` header: `Bearer <token>`
    
    ## Response Format
    All responses follow a consistent format:
    - **Success**: `{ "status": "success", "data": {...} }`
    - **Error**: `{ "status": "error", "message": "...", "code": "..." }`
  contact:
    name: CardioVoice Team
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.cardiovoice.com
    description: Production server
...
paths:
...
  /api/v1/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account and receive an access token immediately.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
            example:
              email: "user@example.com"
              password: "SecurePass123"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                status: success
                message: Registration successful
                data:
                  access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  token_type: bearer
                  expires_in: 86400
                  user:
                    id: "b3428e61-b82a-41a0-a084-5b18af6e472b"
                    email: "user@example.com"
                    created_at: "2025-12-20T10:00:00Z"
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_email:
                  summary: Invalid email format
                  value:
                    status: error
                    message: "Invalid email format"
                    code: "VALIDATION_ERROR"
                weak_password:
                  summary: Password too weak
                  value:
                    status: error
                    message: "Password must be at least 8 characters"
                    code: "VALIDATION_ERROR"
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: error
                message: "Email already registered"
                code: "EMAIL_EXISTS"
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: |
        Authenticate a user and receive an access token.
        
        **Rate Limit**: 5 requests per minute (brute-force protection)
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
            example:
              email: "user@example.com"
              password: "SecurePass123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                status: success
                message: Login successful
                data:
                  access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  token_type: bearer
                  expires_in: 86400
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: error
                message: "Invalid email or password"
                code: "AUTH_FAILED"
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: error
                message: "Rate limit exceeded. Try again later."
                code: "RATE_LIMIT"
  /api/v1/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Retrieve the authenticated user's profile information.
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/User'
              example:
                status: success
                data:
                  id: "b3428e61-b82a-41a0-a084-5b18af6e472b"
                  email: "user@example.com"
                  is_active: true
                  created_at: "2025-12-20T10:00:00Z"
                  updated_at: "2025-12-20T10:00:00Z"
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_token:
                  summary: No token provided
                  value:
                    status: error
                    message: "Authentication required. Please provide a valid token."
                    code: "AUTH_REQUIRED"
                invalid_token:
                  summary: Invalid or expired token
                  value:
                    status: error
                    message: "Invalid or expired token. Please login again."
                    code: "INVALID_TOKEN"
  /api/v1/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: |
        Exchange a valid refresh token for a new access token and refresh token.
        
        **Token Rotation**: The old refresh token is immediately invalidated.
        This is a security feature - each refresh token can only be used once.
        
        **Theft Detection**: If a revoked refresh token is used, all tokens 
        in the same family are revoked (the attacker and legitimate user 
        are both logged out).
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  description: The refresh token from login/register or previous refresh
            example:
              refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
              example:
                status: success
                message: "Token refreshed successfully"
                data:
                  access_token: "new_access_token_here"
                  refresh_token: "new_refresh_token_here"
                  token_type: bearer
                  access_token_expires_in: 900
                  refresh_token_expires_in: 604800
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                expired:
                  summary: Token expired
                  value:
                    status: error
                    message: "Refresh token has expired"
                    code: "REFRESH_FAILED"
                revoked:
                  summary: Token revoked
                  value:
                    status: error
                    message: "Token has been revoked. Please login again."
                    code: "REFRESH_FAILED"
  /api/v1/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout (single device)
      description: |
        Revoke a specific refresh token to logout from a single device.
        
        The access token will remain valid until it expires (15 minutes),
        but the refresh token cannot be used to obtain new tokens.
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  description: The refresh token to revoke
            example:
              refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: "Logged out successfully"
        '400':
          description: Logout failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout-all:
    post:
      tags:
        - Authentication
      summary: Logout from all devices
      description: |
        Revoke all refresh tokens for the authenticated user.
        
        This effectively logs out the user from all devices.
        **Requires authentication**.
      operationId: logoutAll
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logged out from all devices
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: "Logged out from 3 device(s)"
                  data:
                    type: object
                    properties:
                      revoked_tokens:
                        type: integer
                        example: 3
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================
  # ANALYSIS ENDPOINTS
  # ============================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT authentication. Obtain a token via `/api/v1/auth/login`.
        
        Include in header: `Authorization: Bearer <token>`
  schemas:
...
    UserRegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          minLength: 5
          maxLength: 255
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          description: "Must contain at least one letter and one digit"
          example: "SecurePass123"
    UserLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "SecurePass123"
...
    AuthResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: "Login successful"
        data:
          type: object
          properties:
            access_token:
              type: string
              description: JWT access token (short-lived, 15 min)
            refresh_token:
              type: string
              description: Refresh token (long-lived, 7 days)
            token_type:
              type: string
              example: bearer
            access_token_expires_in:
              type: integer
              description: Access token lifetime in seconds
              example: 900
            refresh_token_expires_in:
              type: integer
              description: Refresh token lifetime in seconds
              example: 604800
            user:
              $ref: '#/components/schemas/User'
...